+++ .gitignore
@@ -397,3 +397,4 @@ FodyWeavers.xsd
 # publicized DSP assembly
 lib/Assembly-CSharp-publicized.dll
 /DefaultPath.props
+/lib/Newtonsoft.Json.dll.old
