+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+- v0.8.7
+  + 调整 生物塑料 配方，增加产出
+    
+  + 调整 有深层熔岩星球的钛储量
+
+  + 修复 部分科技不能研究站研究的问题，部分科技研究时间过长的问题
+
+  + 调整 轨道熔炼站为x16，15MW
+        
+  + 修复 mod的科技设定会串存档的问题
+
+  <details>
+	<summary>点击展开历史修改 | Click to view all </summary>
+
+- v0.8.6
+  + 调整 部分科技前置，研究时间，描述
+    
+  + 调整 电路板 配方，以及新增两个电路板配方
+
+  + 修复 读取存档时报错的问题
+
 - v0.8.5
     
   + MOD更名为星环MOD，不兼容旧版本存档，后续还有类月球卫星需要新档生效的更新，因此本版本不建议正式开档
@@ -23,9 +44,6 @@
   + 新增 弹射器增加堆叠发射，由科技【电磁霰射】解锁
     
   + 调整 近防电浆炮 不再可以使用反物质轨道弹，解锁科技独立出来
-
-  <details>
-	<summary>点击展开历史修改 | Click to view all </summary>
   
 - v0.8.4
   + 调整 黑雾单位掉落沙土清零，黑雾建筑掉落沙土为原来的5倍
