+++ data/items_mod.json
@@ -1380,9 +1380,6 @@
         "Type": 3,
         "PreTechOverride": 0,
         "DescFields": [
-            2,
-            3,
-            31,
             1,
             40
         ],
