+++ data/prefabDescs.json
@@ -303,8 +303,8 @@
         "isStation": false,
         "assemblerRecipeType": 1,
         "idleEnergyPerTick": 20000,
-        "workEnergyPerTick": 200000,
-        "assemblerSpeed": 80000,
+        "workEnergyPerTick": 250000,
+        "assemblerSpeed": 160000,
         "stationMaxEnergyAcc": 0,
         "stationMaxItemCount": 0,
         "stationMaxItemKinds": 0,
