+++ data/strings_mod.json
@@ -1,29 +1,29 @@
 ﻿[
   {
-    "Name": "Cosmogenesis",
-    "ZHCN": "宇宙创生",
-    "ENUS": "Cosmogenesis"
+    "Name": "OrbitalRing",
+    "ZHCN": "星环",
+    "ENUS": "OrbitalRing"
   },
   {
-    "Name": "GenesisBookLoadTitle",
-    "ZHCN": "创世之书提示",
-    "ENUS": "From GenesisBook mod"
+    "Name": "OrbitalRingTitle",
+    "ZHCN": "星环提示",
+    "ENUS": "From OrbitalRing mod"
   },
   {
     "Name": "PreloaderNotInstalled",
-    "ZHCN": "⚠ 未检测到创世之书patch补丁！请检查您是否正确安装了创世之书mod。",
-    "ENUS": "⚠ GenesisBook Preloader patch has not been detected, please check your installation, make sure you followed installation instructions correctly."
+    "ZHCN": "⚠ 未检测到星环patch补丁！请检查您是否正确安装了星环mod。",
+    "ENUS": "⚠ OrbitalRing Preloader patch has not been detected, please check your installation, make sure you followed installation instructions correctly."
   },
   {
-    "Name": "ProjectGenesisNotLoaded",
-    "ZHCN": "⚠ 创世之书mod加载失败！请检查您是否正确安装了创世之书mod。",
-    "ENUS": "⚠ GenesisBook has not been loaded, please check your installation, make sure you followed installation instructions correctly."
+    "Name": "ProjectOrbitalRingNotLoaded",
+    "ZHCN": "⚠ 星环mod加载失败！请检查您是否正确安装了星环mod。",
+    "ENUS": "⚠ OrbitalRing has not been loaded, please check your installation, make sure you followed installation instructions correctly."
   },
-    {
-        "Name": "GenesisBookLoadMessage",
-        "ZHCN": "⚠警告，本MOD为以创世之书为基础修改的MOD，还未完成，若您不以测试为目标，请立刻退出。若您在主界面中看到了创世之书的图标、无红字报错，且科技树与原版不同，说明MOD已成功加载，图标还未替换完成，看到创世之书是正常现象。\n请注意，本MOD修改了大量原版内容，因此禁用了上传银河系的功能。\n如果您在游玩MOD时遇到了任何问题，或者有宝贵的意见或建议，欢迎加入MOD交流群",
-        "ENUS": "⚠Warning, this MOD is modified based on the Book of Genesis and is not yet completed. If you do not intend to test it, please exit immediately. If you see the icon of the Book of Genesis in the main interface, with no red error message, and the technology tree is different from the original version, it means that the MOD has been successfully loaded and the icon has not been replaced yet. It is normal to see the Book of Genesis. \nPlease note that this mode modifies a lot of content, and you cannot upload the data the galaxy. \nIf have any issues or ideas about the mod, please join our community on Discord."
-    },
+  {
+    "Name": "OrbitalRingLoadMessage",
+    "ZHCN": "⚠警告，本MOD为以创世之书为基础修改的MOD，还未完成，若您不以测试为目标，请立刻退出。\n若您在主界面中看到了创世之书的图标、无红字报错，且科技树与原版不同，说明MOD已成功加载，图标还未替换完成，看到创世之书是正常现象。\n请注意，本MOD修改了大量原版内容，因此禁用了上传银河系的功能。\n如果您在游玩MOD时遇到了任何问题，或者有宝贵的意见或建议，欢迎加入MOD交流群",
+    "ENUS": "⚠Warning, this MOD is modified based on the Book of Genesis and is not yet completed. If you do not intend to test it, please exit immediately. If you see the icon of the Book of Genesis in the main interface, with no red error message, and the technology tree is different from the original version, it means that the MOD has been successfully loaded and the icon has not been replaced yet. It is normal to see the Book of Genesis. \nPlease note that this mode modifies a lot of content, and you cannot upload the data the galaxy. \nIf have any issues or ideas about the mod, please join our community on Discord."
+  },
   {
     "Name": "BepinExVersionNotMatch",
     "ZHCN": "⚠ BepinEx版本不为5.4.17！这会导致许多不可预知的Bug。",
@@ -50,9 +50,9 @@
     "ENUS": "Update Log"
   },
   {
-    "Name": "创世之书链接",
-    "ZHCN": "https://jq.qq.com/?_wv=1027&k=5bnaDEp3",
-    "ENUS": "https://discord.gg/QwMEeaRZZR"
+    "Name": "星环链接",
+    "ZHCN": "http://qm.qq.com/cgi-bin/qm/qr?_wv=1027&k=5wX2tORAhzm8M76UXKq3eVHTTSQFCIrN&authKey=AgakGrvJByj9BzxUG7lv83OtJaPLxpI8rINNWee1%2BDvh8KTTTckaP57vCWR1CCqt&noverify=0&group_code=1035445959",
+    "ENUS": "http://qm.qq.com/cgi-bin/qm/qr?_wv=1027&k=5wX2tORAhzm8M76UXKq3eVHTTSQFCIrN&authKey=AgakGrvJByj9BzxUG7lv83OtJaPLxpI8rINNWee1%2BDvh8KTTTckaP57vCWR1CCqt&noverify=0&group_code=1035445959"
   },
   {
     "Name": "恢复默认",
