+++ src/Patches/Logic/EjectorPatches.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Reflection.Emit;
 using System.Text;
 using System.Threading.Tasks;
-using static ProjectOrbitalRing.Patches.Logic.ModifyUpgradeTech.ModifyUpgradeTech;
+using static ProjectOrbitalRing.Patches.Logic.ModifyUpgradeTech.Unlock_Save_Load;
 
 namespace ProjectOrbitalRing.Patches.Logic
 {
